refactored file names to fit convention
