+++ main_display/src/components/Counter.jsx
