+++ main_display/src/components/Counters.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Counter from "./counter";
+import Counter from "./Counter";
 
 class Counters extends React.Component {
     state = {
